Refactor API differences into a separate class.
We now have an IAPI interface which defines methods to convert to/from
UUIDs to their REST representations, and to calculate the API-homed full
URL path to a resource.  Add implementations API30 and API31 to handle
the two different implementations so far.  This also simplifies the
various path_to() calls.

Also: Add support for diff_cover to tox.ini to check that all
differences against the master branch have full test coverage.