Fix the template name of the welcome message.
We did not previously complete the renaming of the templates.  Rename:

* user:ack:welcome -> list:user:notice:welcome
* user:ack:goodbye -> list:user:notice:goodbye

The new names are what are actually used in send_welcome_message() and
send_goodbye_message() so the old names were actually never used.

NOTE: If you added welcome or goodbye templates through REST from git head,
the old names will not be used.  Given that this wasn't a feature in 3.0, and
the above use case is very unlikely, I've decided not to do a migration.  If
anybody is affected, the changes can be made manually in withlist.

Closes: #337