Set'ing a template clears its cache.
Setting a template via REST API (or otherwise) means that the local cache for
the template is now obsolete and should be flushed out.

This commit also changes the evict() API a little to evict based on the key
instead of the file-id (which is a hash of the key).