Catch ChildProcessError for failing CI jobs.
We are seeing a lot of failed CI jobs due to transient ChildProcessError which
in my opinion are caused due to a race condition in our logic of terminating
processes that do not behave properly.

The bad behavior seems to stem from a TOCTTOU bug, where when we check for a
process to have died, they seem to be alive, but when we try to kill them, they
have already died at that point.

I don't know if there is a better way to fix that problem, but, just to make
sure that we don't keep failing CI jobs because of it, we are going to catch the
exception raised and just return the function since the job of the function (to
kill the child process) is done.