Fixed a sort-of race condition causing repeated test failures.
This was a very interesting failure, which I think should now be fixed. The
problem seemed to be that since the example plugin is initialized in the
memory and there two worker threads for gunicorn, each of them stored the data
separately with different values.

So, depending on which one served your request, you could get different values
causing test failures when the response was served by a different
worker. Setting the worker threads to 1 during testing should fix this.