Handle TOCTOU error with two subscription requests of same email.
This happens when the same address is being subscribed simultaneous and each of
them pass the "Address does not exist, so let's create it" test and when both
of them try to create the same address, one of them is going to fail as
database would reject the duplicate email address due to unique constraint.

It is hard to recreate the exact conditions, hence there isn't any test, but
since the exception is raised during the SubscriptionWorkflow run, we catch
blanket Exception and return a 400 bad request with the exception message.