Improve the lookup of Members.
Upon investigating reasons for why the Members API is so slow, it was found
that when trying to get a User associated with a Member, the Member model would
do a database query every time, and not caching the result.

So, anytime someone runs `member.user` or `member.display_name` or any
attribute that accesses's member.user, that results in a database lookup if the
Member was subscribed with their address. Subscribing via address is also the
most common case and hence the slow response form the API server.

Looking closely at the implementation of `user_manager.get_user` it just looks
up the Address that you just passed and returns the addres' user.