Reduce the number of database queries in UserManager.
Thanks to Daniel Silverstone for the suggestion and patch.

This reduces the number of queries made for get_user and get_address methods
and uses .one_or_none() instead of calling .count() and .one() separately which
results in two database lookups.

The performance improvement of this change is small, but does reduce the code a
bit too.