Optimize MembershipManager.memberships_pending_removal()
This is the same fix as ee8dfef5ba1c basically.

MembershipManager.memberships_pending_removal() makes a query for a set
of members that potentially meet some bounce criteria and then for each
member found, it does further processing in Python.

While the first query is pretty fast, the function then makes a further
query for each member being checked to look up member.mailing_list. We
can easily optimize this by selecting the interval in the original query
since the mailinglist table is already part of the JOIN.

Signed-off-by: Kunal Mehta <legoktm@debian.org>