Only remove or send warnings to list members when processing bounces
delete_member() expects that it is only operating on members with a role
of "member", and will throw an exception if it can't find it. This mostly
makes sense since we can only unsubscribe members and don't want to
automatically remove list moderators/owners.

This exception would bubble up, triggering a rollback and crashing
the bounce runner (both separate issues).

We apply the same logic to the warning notifications as well since
they'll only be unsubscribed if they have a role of "member".

Fixes #910.