Switch autoresponse_text fields to Text
The autoresponse_text fields are too small in MySQL: only VARCHAR(255).

Make them bigger following the same treatment that info got in 01e871cb5
by switching them to the SAText type.

While we're at it, change SAText to use the utf8mb4 collation on MySQL.
I'm unsure why it was skipped in 10f6f8976, but I don't see any reason
not to do it now.

Fixes #925 and #891.