Improve ARC documentation
I had a bit of a hard time setting up ARC because of gaps in the documentation.

First, I had to hunt down the Unparsable private key exception.

Then I got a weird Syntax Error exception caused by the lack of `authserv_id` in my config. I did not set that initially because as I read the documentation, I thought I only needed that if mailman does **not** do its own DMARC/DKIM checks (the heading "trusted domains" confused me. It makes somewhat sense in schema.cfg, but not in the rendered docs).

I hope this is an improvement, this wording and layout would've helped me a lot anyway.