Send email notifications UTF-8 encoded
The list:admin:action:subscribe and list:admin:action:unsubscribe templates include username, which username includes the displayname and email address.  The displayname shall be sent per email in the template to the list owners RFC2047-decoded.  This implies, that the email must be sometimes UTF-8 encoded, or at least not us-ascii, so that it can contain correctly the displayname.

This change:
* coverts the emails from UserNotification to UTF-8, if the preferred charset cannot include the input correctly.
* in the list:admin:action:subscribe and list:admin:action:unsubscribe templates puts the RFC2047-decoded displayname in username.